move write stats to it's own file
